+++ src/main/java/school/faang/BJS2_43751/Main.java
@@ -1,0 +1,9 @@
+package school.faang.BJS2_43751;
+
+public class Main {
+    public static void main(String[] args) {
+        new User("Bob", 25, "Google", "London");
+        new User("Bob", 25, "Google", "");
+    }
+
+}
